requirements.txt
Flask
Flask-CORS
requests
zipfile36
Pillow
google-cloud-aiplatform 